# docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Personal Style Network API",
    "version": "1.0.0",
    "description": "API cho ứng dụng Personal Style Network",
    "contact": {
      "name": "PSN Team",
      "email": "support@psn.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3052/v1/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/users/profile": {
      "get": {
        "summary": "Lấy thông tin profile của người dùng hiện tại",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin profile người dùng"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      },
      "put": {
        "summary": "Cập nhật thông tin profile người dùng",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bodyType": {
                    "type": "string",
                    "enum": [
                      "hourglass",
                      "pear",
                      "apple",
                      "rectangle",
                      "inverted_triangle"
                    ]
                  },
                  "seasonPreferences": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "spring",
                        "summer",
                        "fall",
                        "winter"
                      ]
                    }
                  },
                  "favoriteColors": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "stylePreferences": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "measurements": {
                    "type": "object",
                    "properties": {
                      "height": {
                        "type": "number"
                      },
                      "weight": {
                        "type": "number"
                      },
                      "bust": {
                        "type": "number"
                      },
                      "waist": {
                        "type": "number"
                      },
                      "hips": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật profile thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Lấy thông tin người dùng theo ID",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của người dùng"
          }
        ],
        "responses": {
          "200": {
            "description": "Thông tin người dùng"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        }
      },
      "put": {
        "summary": "Cập nhật thông tin người dùng",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của người dùng"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật thông tin người dùng thành công"
          },
          "400": {
            "description": "Dữ liệu không hợp lệ"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        }
      },
      "delete": {
        "summary": "Xóa người dùng (chỉ admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của người dùng"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa người dùng thành công"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "403": {
            "description": "Không có quyền truy cập"
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Lấy danh sách người dùng (chỉ admin)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Số lượng mỗi trang"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "default": "ctime"
            },
            "description": "Sắp xếp theo (ctime, mtime)"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách người dùng"
          },
          "401": {
            "description": "Không được ủy quyền"
          },
          "403": {
            "description": "Không có quyền truy cập"
          }
        }
      }
    },
...
  },
...
}